Add signature version
Signature are now versioned. This means their serialisation, or
signature algorithm can be tuned down the line.

This commit proposes two versions:
* 0x0001 = protobuf and ed25519
* 0x0002 = bincode and ed25519

Version 0x0002 is the closest to the existing.

The introduction of the version field is not backward compatible.
Given the early stage of deployment, it's better to break now, rather
than introduce a default non-versioned signature.